Split mixed SPC and repository sync tests #295
